docs: describe how to use podman with kind
This patch adds a section to the README describing how to use
Podman with `cloud-provider-kind`. Podman does not allow sharing
the host networking or binding to privileged ports so the
`-enable-lb-port-mapping` option can be used to get an ephemeral port on
the host that maps to the `LoadBalancer` listening port.