fix(tunnel): propagate half-closed TCP connections
This change ensures that tunneled TCP connections propagate shut downs
when they occur, allowing the opposite party to be signaled
appropriately.

A TCP connection can be "half-closed" when `A` shuts down `A->B` by
sending a FIN to `B`, and `B` responds with an ACK. The connection
becomes "fully-closed" when the same procedure is followed for the
`A<-B` direction. When a TCP connection is half-closed in the `A->B` direction,
data may continue to flow in the `A<-B` direction.

This signaling exists as a basic function of the TCP protocol, allowing either
party to indicate that they are done sending data, and that their peer can
finish up as they please. Data can continue to flow in the opposite direction
until they are ready to shut down their side of the connection.

When cloud-provider-kind (`t`) acts as a tunnel for TCP connections
between `A` and `B`, we have two connections `A-t` and `t-B`, and four
flows:

- `A->t` and `t->B` (representing `A->B`).
- `A<-t` and `t<-B` (representing `A<-B`).

Prior to this change, when `A` shut down `A->t`, the tunnel would not
shut down `t->B`, leaving `A-t` half-closed and `B-t` not closed in any
way. The same is true for the opposite direction, as well.
obviously prevents `B` from knowing

After this change, when `A` shuts down `A->t`, cloud-provider-kind
immediately shuts down `t->B`. The same is true when `B` shuts down
`t<-B`, `A<-t` is immediately shut down. Flow directions are maintained
independently from one-another.

Note: this change effectively restricts tunneled connections to
only supporting `udp`, `tcp`, `tcp4`, and `tcp6` protocols. As far as I can
tell, docker containers will not expose other protocol types (like unix
domain sockets).