Move Gateway API section in README
The Gateway API support section was added inside the "How to use it"
section of the README, disrupting the flow and making it look like
things like "Running the provider" were subsections of Gateway API
rather than initial setup instructions.

This moves the Gateway API support section just prior to "How to use it"
to keep the flow of the setup instructions.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>