Add contextual logging to controllers
This adds contextual logging only to our controller classes to help
differentiate the logging coming from each. For now it is limited to the
`controller.go` files since those are the key conflicting filenames in
the log output. To switch entirely over to contextual logging as a whole
will require more intrusive changes, so this just addresses are major
pain point for now.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>